Introduce ngrow in the EB generation step
When construction the level set information for embedded boundaries
we assumed that a ghost cell width of 4 would be enough.
This is wrong since we can have arbitrary many ghost cells now.
This commit introduces a ngrow parameter which controls the amount of
ghost cells for embedded boundary geometry information storage.